requirements.txt
python==3.9.21

h5py==3.13.0
matplotlib==3.9.4
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
pillow==11.2.1
scikit-learn==1.6.1
seaborn==0.13.2
shap==0.47.2
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
torchxrayvision==1.3.5
tqdm==4.67.1
transformers==4.34.0